Split odin_controller and rename classes
