+++ src/odin_fastcs/frame_receiver.py
@@ -1,0 +1,5 @@
+from odin_fastcs.odin_controller import OdinAdapterController
+
+
+class FrameRecieverController(OdinAdapterController):
+    pass
