+++ tests/test_controllers.py
@@ -4,12 +4,12 @@ import pytest
 from fastcs.attributes import AttrR, AttrRW
 from fastcs.datatypes import Bool, Float, Int
 
-from odin_fastcs.http_connection import HTTPConnection
-from odin_fastcs.odin_controller import (
-    OdinFPController,
-    OdinFPPluginController,
-    OdinSubController,
+from odin_fastcs.frame_processor import (
+    FrameProcessorController,
+    FrameProcessorPluginController,
 )
+from odin_fastcs.http_connection import HTTPConnection
+from odin_fastcs.odin_controller import OdinAdapterController
 from odin_fastcs.util import OdinParameter
 
 HERE = Path(__file__).parent
@@ -21,12 +21,12 @@ def test_create_attributes():
         OdinParameter(uri=["write_bool"], metadata={"type": "bool", "writeable": True}),
         OdinParameter(uri=["group", "float"], metadata={"type": "float"}),
     ]
-    controller = OdinSubController(HTTPConnection("", 0), parameters, "api/0.1")
+    controller = OdinAdapterController(HTTPConnection("", 0), parameters, "api/0.1")
 
     controller._create_attributes()
 
     match controller:
-        case OdinSubController(
+        case OdinAdapterController(
             read_int=AttrR(datatype=Int()),
             write_bool=AttrRW(datatype=Bool()),
             group_float=AttrR(datatype=Float(), group="Group"),
@@ -42,7 +42,7 @@ def test_fp_process_parameters():
         OdinParameter(["0", "config", "hdf", "frames"], metadata={}),
     ]
 
-    fpc = OdinFPController(HTTPConnection("", 0), parameters, "api/0.1")
+    fpc = FrameProcessorController(HTTPConnection("", 0), parameters, "api/0.1")
 
     fpc._process_parameters()
     assert fpc._parameters == [
@@ -72,7 +72,7 @@ async def test_fp_create_plugin_sub_controllers():
         ),
     ]
 
-    fpc = OdinFPController(HTTPConnection("", 0), parameters, "api/0.1")
+    fpc = FrameProcessorController(HTTPConnection("", 0), parameters, "api/0.1")
 
     await fpc._create_plugin_sub_controllers(["hdf"])
 
@@ -86,7 +86,7 @@ async def test_fp_create_plugin_sub_controllers():
     ]
     match fpc.get_sub_controllers():
         case {
-            "HDF": OdinFPPluginController(
+            "HDF": FrameProcessorPluginController(
                 _parameters=[
                     OdinParameter(
                         uri=["status", "hdf", "frames_written"],
